finalizing submodulification by making idioms.py and switching interior module to composite
